+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- semver environment variables
+
 ## [1.1.0] - 2019-07-14
 ### Added
 - logrus hook to store recent system logs